注册业务实现，上传文件细节修改

+++ zx-platform/zx-bean/src/main/java/com/zx/share/platform/bean/zx/ZxFileManagerCDE.java
@@ -56,7 +56,7 @@ public class ZxFileManagerCDE extends IdEntity{
 	private Date createTime;
 	
 	@Column(name = "create_id")
-	private int createId;
+	private Long createId;
 	
 	@Column(name = "update_time")
 	private Date updateTime;
